+++ tests/commands/test_set_closed_codelist_enums.py
@@ -4,6 +4,8 @@ from io import StringIO
 from tempfile import TemporaryDirectory
 from unittest.mock import patch
 
+import pytest
+
 from ocdskit.cli.__main__ import main
 
 schema = '''{
@@ -46,7 +48,7 @@ schema = '''{
         "string",
         "null"
       ],
-      "codelist": "e.csv",
+      "codelist": "a.csv",
       "openCodelist": true
     }
   }
@@ -106,7 +108,7 @@ schema_with_enum = '''{
         "string",
         "null"
       ],
-      "codelist": "e.csv",
+      "codelist": "a.csv",
       "openCodelist": true
     }
   }
@@ -118,12 +120,32 @@ codelist = 'Code\nfoo\nbar\n'
 
 def test_command(monkeypatch):
     with TemporaryDirectory() as d:
-        os.mkdir(os.path.join(d, 'codelists'))
-
         with open(os.path.join(d, 'release-schema.json'), 'w') as f:
             f.write(schema)
 
-        for basename in ('a', 'b', 'c', 'd'):
+        for directory in ('codelists', 'codelists_translated'):
+            os.mkdir(os.path.join(d, directory))
+            for basename in ('a', 'b', 'c', 'd'):
+                with open(os.path.join(d, directory, '{}.csv'.format(basename)), 'w') as f:
+                    f.write(codelist)
+
+        with patch('sys.stdout', new_callable=StringIO) as actual:
+            monkeypatch.setattr(sys, 'argv', ['ocdskit', 'set-closed-codelist-enums', d])
+            main()
+
+        assert actual.getvalue() == ''
+
+        with open(os.path.join(d, 'release-schema.json')) as f:
+            assert f.read() == schema_with_enum
+
+
+def test_unused_codelists(monkeypatch, caplog):
+    with TemporaryDirectory() as d:
+        with open(os.path.join(d, 'release-schema.json'), 'w') as f:
+            f.write(schema)
+
+        os.mkdir(os.path.join(d, 'codelists'))
+        for basename in ('a', 'b', 'c', 'd', 'e'):
             with open(os.path.join(d, 'codelists', '{}.csv'.format(basename)), 'w') as f:
                 f.write(codelist)
 
@@ -136,10 +158,57 @@ def test_command(monkeypatch):
         with open(os.path.join(d, 'release-schema.json')) as f:
             assert f.read() == schema_with_enum
 
+        assert len(caplog.records()) == 1
+        assert caplog.records()[0].levelname == 'ERROR'
+        assert caplog.records()[0].message == 'unused codelists: e.csv'
 
-def test_conflicting_codelists(monkeypatch):
-    pass
+
+def test_missing_codelists(monkeypatch, caplog):
+    with TemporaryDirectory() as d:
+        with open(os.path.join(d, 'release-schema.json'), 'w') as f:
+            f.write(schema)
+
+        os.mkdir(os.path.join(d, 'codelists'))
+
+        with pytest.raises(KeyError) as excinfo:
+            with patch('sys.stdout', new_callable=StringIO) as actual:
+                monkeypatch.setattr(sys, 'argv', ['ocdskit', 'set-closed-codelist-enums', d])
+                main()
+
+        assert actual.getvalue() == ''
+
+        with open(os.path.join(d, 'release-schema.json')) as f:
+            assert f.read() == schema
+
+        assert len(caplog.records()) == 0
+        assert excinfo.value.args == ('a.csv',)
 
 
-def test_unused_codelists(monkeypatch):
-    pass
+def test_conflicting_codelists(monkeypatch, caplog):
+    with TemporaryDirectory() as d:
+        with TemporaryDirectory() as e:
+            for directory in (d, e):
+                os.mkdir(os.path.join(directory, 'codelists'))
+
+                with open(os.path.join(directory, 'release-schema.json'), 'w') as f:
+                    f.write(schema)
+
+            for basename in ('a', 'b', 'c', 'd'):
+                with open(os.path.join(d, 'codelists', '{}.csv'.format(basename)), 'w') as f:
+                    f.write(codelist)
+
+            with open(os.path.join(e, 'codelists', 'a.csv'), 'w') as f:
+                f.write('Code\nbaz\n')
+
+            with patch('sys.stdout', new_callable=StringIO) as actual:
+                monkeypatch.setattr(sys, 'argv', ['ocdskit', 'set-closed-codelist-enums', d, e])
+                main()
+
+            assert actual.getvalue() == ''
+
+            with open(os.path.join(e, 'release-schema.json')) as f:
+                assert f.read() == schema_with_enum
+
+            assert len(caplog.records()) == 1
+            assert caplog.records()[0].levelname == 'ERROR'
+            assert caplog.records()[0].message == 'conflicting codelists: a.csv'